AI/requirements.txt
fastapi
uvicorn
langchain
qdrant-client
sentence-transformers
pypdf
python-dotenv
openai
langchain-openai
langchain-groq
langchain-huggingface
langchain-qdrant
